==> 1_main.Rmd <==
---
title: "Data exploration"
---

```{r setup, warning = F, message = F}
if (!require("dplyr")) install.packages("dplyr")
if (!require("tidyr")) install.packages("tidyr")
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("foreign")) install.packages("foreign")
if (!require("openxlsx")) install.packages("openxlsx")

library(dplyr)
library(tidyr)
library(ggplot2)
library(foreign)
library(openxlsx)

setwd("/Users/jlein/Downloads/thesis/2_data")

rm(list = ls())
```

## Merge with macro data
```{r warning = F, message = F, results = 'hide'}
# GDP growth
macro_data_1 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_gdpgrowth")
macro_data_1 <- macro_data_1 %>% fill(country, .direction = "down") %>% mutate(year = as.character(year))

# Unemployment rates
macro_data_21 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_unemp_total")
macro_data_21 <- macro_data_21 %>% rename("Czech Republic" = "Czech.Republic")%>% pivot_longer(!year, names_to = "country", values_to = "macro_unemp_total") %>% arrange(country, macro_unemp_total) %>% mutate(year = as.character(year))

macro_data_22 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_unemp_female")
macro_data_22 <- macro_data_22 %>% rename("Czech Republic" = "Czech.Republic")%>% pivot_longer(!year, names_to = "country", values_to = "macro_unemp_gender") %>% arrange(country, macro_unemp_gender) %>% mutate(gender = "Female") %>% mutate(year = as.character(year))

macro_data_23 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_unemp_male")
macro_data_23 <- macro_data_23 %>% rename("Czech Republic" = "Czech.Republic")%>% pivot_longer(!year, names_to = "country", values_to = "macro_unemp_gender") %>% arrange(country, macro_unemp_gender) %>% mutate(gender = "Male") %>% mutate(year = as.character(year))

macro_data_2 <- rbind(macro_data_22, macro_data_23)

rm(macro_data_22, macro_data_23)

# Life expectancy at age 65
macro_data_31 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_lifeexp65_female", colNames = F)
macro_data_31 <- macro_data_31 %>% 
  slice(3:n()) %>%
  rename("country" = "X1", "year" = "X2", "macro_lifeexp65_gender" = "X3") %>%
  fill(country, .direction = "down") %>% 
  mutate(gender = "Female") %>% mutate(year = as.character(year))

macro_data_32 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_lifeexp65_male", colNames = F)
macro_data_32 <- macro_data_32 %>% 
  slice(3:n()) %>%
  rename("country" = "X1", "year" = "X2", "macro_lifeexp65_gender" = "X3") %>%
  fill(country, .direction = "down") %>% 
  mutate(gender = "Male") %>% mutate(year = as.character(year))

macro_data_3 <- rbind(macro_data_31, macro_data_32)

rm(macro_data_31, macro_data_32)

# Effective retirement ages
macro_data_41 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_effretage_female")
macro_data_41 <- macro_data_41 %>% rename("Czech Republic" = "Czech.Republic") %>% pivot_longer(!year, names_to = "country", values_to = "macro_effretage_gender") %>% arrange(country, macro_effretage_gender) %>% mutate(gender = "Female") %>% mutate(year = as.character(year))

macro_data_42 <- read.xlsx("data_other/macro_data.xlsx", sheet = "macro_effretage_male")
macro_data_42 <- macro_data_42 %>% rename("Czech Republic" = "Czech.Republic") %>% pivot_longer(!year, names_to = "country", values_to = "macro_effretage_gender") %>% arrange(country, macro_effretage_gender) %>% mutate(gender = "Male") %>% mutate(year = as.character(year))

macro_data_4 <- rbind(macro_data_41, macro_data_42)

rm(macro_data_41, macro_data_42)
```

```{r warning = F, message = F, results = 'hide'}
job_episodes <- read.dta("data_share/sharewX_rel7-1-0_gv_job_episodes_panel_stata/sharewX_rel7-1-0_gv_job_episodes_panel.dta")

job_episodes <- job_episodes %>% 
  select(mergeid, country, year, yrbirth, gender, age, ret_age, nchildren, withpartner, married, ordjob, job_title) %>%
  rename("macro_retage_gender" = "ret_age") %>%
  group_by(mergeid) %>%
  fill(macro_retage_gender, .direction = "up")
```

```{r}
data_main <- job_episodes %>% mutate(year = as.character(year)) %>% 
  left_join(macro_data_1, by = c("country", "year")) %>%
  left_join(macro_data_21, by = c("country", "year")) %>%
  left_join(macro_data_2, by = c("country", "year", "gender")) %>%
  left_join(macro_data_3, by = c("country", "year", "gender")) %>%
  left_join(macro_data_4, by = c("country", "year", "gender"))

data_main$macro_gdpgrowth <- round(data_main$macro_gdpgrowth, 2)
data_main$macro_unemp_total <- round(data_main$macro_unemp_total, 2)
data_main$macro_unemp_gender <- round(data_main$macro_unemp_gender, 2)
data_main$macro_lifeexp65_gender <- round(as.numeric(data_main$macro_lifeexp65_gender), 2)
data_main$macro_effretage_gender <- round(data_main$macro_effretage_gender, 2)

rm(macro_data_1, macro_data_2, macro_data_21, macro_data_3, macro_data_4, job_episodes)
```

```{r}
save.image("data_output/data_1_main.RData")
```
