==> 8_pcem.Rmd <==
---
title: "Data analysis"
date: "28/01/2022"
---

```{r setup, warning = F, message = F}
if (!require("dplyr")) install.packages("dplyr")
if (!require("tidyr")) install.packages("tidyr")
if (!require("survival")) install.packages("survival")
if (!require("stargazer")) install.packages("stargazer")
if (!require("modeest")) install.packages("modeest")

library(dplyr)
library(tidyr)
library(survival)
library(stargazer)
library(modeest)

#setwd("/home/jan/Documents/thesis/2_data")

rm(list = ls())

load("data_output/data_final.RData")

#https://data.princeton.edu/wws509/r/recidivism
```

```{r}
data_final <- data_final %>% 
  filter(age >= 50) %>%
  filter(age <= age_retired) %>%
  group_by(mergeid) %>%
  mutate(retired = ifelse(!is.na(age_retired), 1, 0))

data_final$class_oesch8 <- relevel(as.factor(data_final$class_oesch8), ref = 'Production workers')
data_final$family_status <- relevel(as.factor(data_final$family_status), ref = 'married')
data_final$gender <- relevel(as.factor(data_final$gender), ref = 'Male')

# also include those not retiring
```

```{r}
breaks <- c(55, 59, 60, 61, 64, 65)

test <- survSplit(Surv(age, retired) ~ ., data = data_final, cut = breaks, episode = "interval", start = "start")

test$start <- ifelse(test$start == 0, 49, test$start)

test <- mutate(test, exposure = age - start, interval = factor(interval,  labels = paste("[", c(50, breaks), ",", c(breaks,100), "]", sep = "")))

test <- test %>% dplyr::select(mergeid, retired, interval, class_oesch8, gender, ill_health, start, age, age_retired, year, country, dist_retage, exposure, family_status, nchildren, education_years, employed_years) %>% na.omit()

test <- test %>%
  mutate(retired = ifelse(age == age_retired, 1, 0)) %>%
  group_by(mergeid, age) %>%
  filter(year == min(year)) %>%
  group_by(mergeid, interval) %>%
  filter(age == max(age))
```

```{r}
model_1 <- glm(retired ~ interval + class_oesch8 + gender + age + year + country + offset(log(exposure)), data = test, family = poisson(link = "log"))

model_2 <- glm(retired ~ interval + class_oesch8 + gender + ill_health + age + year + dist_retage + country + offset(log(exposure)) + family_status + nchildren + education_years + education_years, data = test, family = poisson(link = "log"))

model_3 <- glm(retired ~ interval + class_oesch8*gender + age + year + country + offset(log(exposure)), data = test, family = poisson(link = "log"))

model_4 <- glm(retired ~ interval + class_oesch8*gender + ill_health + age + year + dist_retage + country + offset(log(exposure)) + family_status + nchildren + education_years + education_years, data = test, family = poisson(link = "log"))

stargazer(model_1, model_2, model_3, model_4, type = "text", omit = c("country", "exposure", "interval", "year", "Constant"), apply.coef=exp, t.auto = F, p.auto = F)

summary(model_4)
```

```{r}
```
