==> 11_prediction.Rmd <==
---
title: "Data analysis"
date: "28/01/2022"
---

```{r setup, warning = F, message = F}
if (!require("dplyr")) install.packages("dplyr")
if (!require("tidyr")) install.packages("tidyr")
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("nnet")) install.packages("nnet")
if (!require("stargazer")) install.packages("stargazer")

library(dplyr)
library(tidyr)
library(ggplot2)
library(nnet)
library(stargazer)
library(ggeffects)
library(emmeans)

rm(list = ls())

load("data_output/data_final.RData")
```

```{r}
formula_full <- as.formula("cluster ~ gender + age + I(age^2) + I(age^3) + ill_health + as.numeric(year_birth)")

model <- multinom(formula_full, data_final, Hess = T)

plot <- ggemmeans(model, terms = c("age [all]", "gender"))
  
plot <- plot %>% filter(response.level != "not_retired")

pdf("outputs/plot_pred_age.pdf", width = 8, height = 4)
ggplot(data = plot %>% filter(x %in% c(50:80)), aes(x = x, y = predicted, color = response.level)) +
  facet_wrap(~ group) +
  geom_line() +
  theme_classic() +
  theme(axis.title.x=element_blank(), axis.title.y=element_blank()) +
  scale_color_brewer(palette = "Set2", name = element_blank()) +
  theme(legend.position = "bottom")
dev.off()
```

```{r}
formula_full <- as.formula("cluster ~ gender + age + ill_health + year_birth")

model <- multinom(formula_full, data_final %>% filter(cluster != "not_retired"), Hess = T)

plot <- ggemmeans(model, terms = c("year_birth", "gender"), condition = c(year_birth = c(1940)))

pdf("outputs/plot_pred_year.pdf", width = 10, height = 5)
ggplot(data = plot, aes(x = x, y = predicted, ymin = conf.low, ymax = conf.high, color = response.level)) +
  facet_wrap(~ group) +
  geom_line() +
  geom_ribbon(alpha = 0.5) +
  theme_classic() +
  theme(axis.title.x=element_blank(), axis.title.y=element_blank()) +
  scale_color_brewer(palette = "Set2", name = element_blank()) +
  theme(legend.position = "bottom")
dev.off()
```

```{r}
```
